Include platform-specific headers to get at internal representation of Bignum
Starting from MRI Ruby 2.2, internal details of Bignum are not exposed in the
main ruby.h header file which C extensions include.

However, our C code needs to poke into those internals. Therefore, if we are compiling
under a version of Ruby which does not provide the needed declarations in ruby.h,
use a platform-specific header file. The declarations in the platform-specific header
will be copied from the Ruby interpreter source for the version of Ruby in use.