Integer#popcount raises error if called on negative integer
Previously, #popcount would take the absolute value of a negative number and
operate on that. This behavior was at least consistent, but it's very unlikely
that it would ever be useful.

Since integers in Ruby don't have any fixed bit width (at least conceptually),
"popcount" is not really a meaningful operation on them. The chances are high that
invocation on a negative integer may be unintended; in such cases, raising an exception
may alert the programmer to a bug.