Add methods to Integer class, not Fixnum and Bignum (works on Ruby 2.4+)
As of Ruby 2.4, there are no Fixnum and Bignum classes. There is just the Integer class
(which existed even before), and Integer objects can internally be of FIXNUM_T or BIGNUM_T
type.

So use macros to generate boilerplate functions to dispatch off the object type and call
one of the Fixnum-or-Bignum-specific functions we had before.

Since the Integer class existed even in earlier versions of Ruby, and was the superclass
of both Fixnums and Bignums, the same code also works for earlier versions. It is actually
a bit cleaner, if anything.